refactor(cli): split command domains from main
Move capture, recall, memory, synthesis, team, and admin commands into
canonical modules. Keep main focused on lifecycle, root wiring, health,
and status behavior while preserving command order and aliases.

Tests now patch each command's owner instead of using the old main module
as a compatibility seam.
